+++ README.md
@@ -1,6 +1,6 @@
 # audio-reative-led
 
 
-run on pi with: `sudo python3 visualization.py <pattern>`
+run on pi with: `sudo python3 run.py <pattern>`
 
 pattern can currently be `spectrum`, `energy`, or `scroll`
